requirements.txt
# Base requirements
langchain
langchain_openai
httpx
python-dotenv
psutil
tenacity

# For testing
pytest
pytest-asyncio
